로또 4단계 구현 - 불필요한 import 제거 - 생성자에 주석 추가 - 수동 입력 없을 경우 오류 발생하지 않도록 수정 - BigDecimal 자료형 도입

+++ src/main/java/lotto/domain/LottoResult.java
@@ -1,5 +1,6 @@
 package lotto.domain;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 /**
@@ -9,9 +10,9 @@ import java.util.List;
  */
 public class LottoResult {
     private final List<String> message;
-    private final float totalRateOfReturn;
+    private final BigDecimal totalRateOfReturn;
 
-    public LottoResult(List<String> message, float totalRateOfReturn) {
+    public LottoResult(List<String> message, BigDecimal totalRateOfReturn) {
         this.message = message;
         this.totalRateOfReturn = totalRateOfReturn;
     }
@@ -20,7 +21,7 @@ public class LottoResult {
         return message;
     }
 
-    public float getTotalRateOfReturn() {
+    public BigDecimal getTotalRateOfReturn() {
         return totalRateOfReturn;
     }
 }
